# tweaks to lossless jpeg recompression