--- a/lib/jxl/enc_frame.cc
+++ b/lib/jxl/enc_frame.cc
@@ -794,10 +794,10 @@ class LossyFrameEncoder {
 
     auto& dct = enc_state_->shared.block_ctx_map.dc_thresholds;
     auto& num_dc_ctxs = enc_state_->shared.block_ctx_map.num_dc_ctxs;
-    enc_state_->shared.block_ctx_map.num_dc_ctxs = 1;
+    num_dc_ctxs = 1;
     for (size_t i = 0; i < 3; i++) {
       dct[i].clear();
-      int num_thresholds = (CeilLog2Nonzero(total_dc[i]) - 10) / 2;
+      int num_thresholds = (CeilLog2Nonzero(total_dc[i]) - 12) / 2;
       // up to 3 buckets per channel:
       // dark/medium/bright, yellow/unsat/blue, green/unsat/red
       num_thresholds = std::min(std::max(num_thresholds, 0), 2);
@@ -823,9 +823,8 @@ class LossyFrameEncoder {
       ctx_map[i] = i / lbuckets;
       // up to 3 contexts for chroma
       ctx_map[kNumOrders * num_dc_ctxs + i] =
-          num_dc_ctxs / lbuckets + (i % lbuckets);
-      ctx_map[2 * kNumOrders * num_dc_ctxs + i] =
-          num_dc_ctxs / lbuckets + (i % lbuckets);
+          ctx_map[2 * kNumOrders * num_dc_ctxs + i] =
+              num_dc_ctxs / lbuckets + (i % lbuckets);
     }
     enc_state_->shared.block_ctx_map.num_ctxs =
         *std::max_element(ctx_map.begin(), ctx_map.end()) + 1;
@@ -836,7 +835,7 @@ class LossyFrameEncoder {
     shared.frame_header.UpdateFlag(false, FrameHeader::kUseDcFrame);
     auto compute_dc_coeffs = [&](int group_index, int /* thread */) {
       modular_frame_encoder->AddVarDCTDC(dc, group_index, /*nl_dc=*/false,
-                                         enc_state_);
+                                         enc_state_, /*jpeg_transcode=*/true);
       modular_frame_encoder->AddACMetadata(group_index, /*jpeg_transcode=*/true,
                                            enc_state_);
     };
@@ -886,6 +885,7 @@ class LossyFrameEncoder {
     RunOnPool(pool_, 0, shared.frame_dim.num_groups, tokenize_group_init,
               tokenize_group, "TokenizeGroup");
     *frame_header = shared.frame_header;
+    doing_jpeg_recompression = true;
     return true;
   }
 
@@ -906,7 +906,7 @@ class LossyFrameEncoder {
                                               writer, kLayerDequantTables,
                                               aux_out_, modular_frame_encoder));
     if (enc_state_->cparams.speed_tier <= SpeedTier::kTortoise) {
-      ClusterGroups(enc_state_);
+      if (!doing_jpeg_recompression) ClusterGroups(enc_state_);
     }
     size_t num_histo_bits =
         CeilLog2Nonzero(enc_state_->shared.frame_dim.num_groups);
@@ -1004,6 +1004,7 @@ class LossyFrameEncoder {
   ThreadPool* pool_;
   AuxOut* aux_out_;
   std::vector<EncCache> group_caches_;
+  bool doing_jpeg_recompression = false;
 };
 
 Status EncodeFrame(const CompressParams& cparams_orig,