--- a/lib/jxl/enc_modular.h
+++ b/lib/jxl/enc_modular.h
@@ -42,7 +42,7 @@ class ModularFrameEncoder {
   // `nl_dc` decides whether to apply a near-lossless processing to the DC or
   // not.
   void AddVarDCTDC(const Image3F& dc, size_t group_index, bool nl_dc,
-                   PassesEncoderState* enc_state);
+                   PassesEncoderState* enc_state, bool jpeg_transcode);
   // Creates a modular image for the AC metadata of the given group
   // (`group_index`).
   void AddACMetadata(size_t group_index, bool jpeg_transcode,