@veluca93 2021-10-01
I'd prefer to keep that in a separate change though :)
I'm not sure it is universally a good idea...

On Fri, 1 Oct 2021, 17:16 Jon Sneyers, ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In lib/jxl/enc_modular.cc
> <https://github.com/libjxl/libjxl/pull/679#discussion_r720334827>:
>
> > @@ -1421,6 +1421,13 @@ void ModularFrameEncoder::AddVarDCTDC(const Image3F& dc, size_t group_index,
>    if (cparams.speed_tier >= SpeedTier::kSquirrel) {
>      stream_options[stream_id].tree_kind = ModularOptions::TreeKind::kWPFixedDC;
>    }
> +  if (cparams.speed_tier < SpeedTier::kSquirrel && !nl_dc) {
>
> Oh you're right, it's false IF we do JPEG recompression but it can also be
> false in other cases (when distance < 2).
>
> So the way I wrote it also makes regular VarDCT spend more time on DC at
> effort 8 and 9. That was not intended, but it's probably not a bad idea,
> actually. At those slow speeds, it's doing a lot of stuff anyway, so
> spending a bit more time on DC is not going to be a huge problem, I think.
>
> Some quick results: (timing is not accurate at all, laptop was running hot
> after doing the 'before')
>
> Before:
>
> Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
> ------------------------------------------------------------------------------------------------------------
> jxl:8           57482 11026557    1.5345979   0.304  19.021   2.80633187   0.62708369  0.962321348304      0
> Aggregate:      57482 11026557    1.5345979   0.304  19.021   2.80633187   0.62708369  0.962321348304      0
>
> After:
>
> Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
> ------------------------------------------------------------------------------------------------------------
> jxl:8           57482 11006224    1.5317681   0.266  15.616   2.80633187   0.62708369  0.960546825216      0
> Aggregate:      57482 11006224    1.5317681   0.266  15.616   2.80633187   0.62708369  0.960546825216      0
>
> So for e8 that's a 0.18% overall density improvement for d1, at the cost
> of slower encode and decode. Not sure if that's worth it, but it may be
> (after all, if you really care about encode/decode speed more than about
> compression, you'll likely not use -e 8 and --faster_decoding 0).
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/libjxl/libjxl/pull/679#discussion_r720334827>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAOPAI2NUEG5CVXB4YOKBKTUEXGFZANCNFSM5FE2DP2A>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>