@jonsneyers 2021-10-01
> I'd prefer to keep that in a separate change though :) I'm not sure it is universally a good idea...

Yes, it was not the intention of this PR to change from-pixels VarDCT, and if we're going to change that, I agree it should be in a separate PR with proper benchmarking of impact on density and speed.