@jonsneyers 2021-10-03
Changed it so the changes in DC encoding apply only to the jpeg recompression case, not the general VarDCT case where it accidentally also applied.